% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.assay_picker.R
\name{assay_picker.ui}
\alias{assay_picker.ui}
\alias{assay_picker.server}
\title{Select assay}
\usage{
assay_picker.ui(id, label = "Assay")

assay_picker.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a way to select an assay to use
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(assay_picker.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(assay_picker.server, id='page_name')
}}

}
