% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boxplot_plot.R
\name{boxplot_plot.ui}
\alias{boxplot_plot.ui}
\alias{boxplot_plot.server}
\title{Make a boxplot of a feature}
\usage{
boxplot_plot.ui(id, feature)

boxplot_plot.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{feature}{name of the column in the Seurat object to plot}
}
\description{
Creates a plot of points and thresholds
}
\examples{

\dontrun{
ui <- fluidPage(boxplot_plot.ui(id='page_name', feature='column_name'))
server <- function(input, output, session) {
  callModule(boxplot_plot.server, id='page_name-column_name')
}} 

}
