% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.cell_filtering.R
\name{cell_filtering.server}
\alias{cell_filtering.server}
\title{React to filtering thresholds and filter Seurat}
\usage{
cell_filtering.server(input, output, session, seurat, ...)
}
\description{
React to filtering thresholds and filter Seurat
}
\details{
Updates the Seurat object and thresholds reactive values.
}
