% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.cluster_id_picker.R
\name{cluster_id_picker.ui}
\alias{cluster_id_picker.ui}
\alias{cluster_id_picker.server}
\title{Pick clusters from selected resolution}
\usage{
cluster_id_picker.ui(id, label = "Cluster selection")

cluster_id_picker.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to select clusters(s)
}
\details{
Updates reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(cluster_id_picker.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(cluster_id_picker.server, id='page_name')
}}

}
