% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.cluster_resolution_picker.R
\name{cluster_resolution_picker.ui}
\alias{cluster_resolution_picker.ui}
\alias{cluster_resolution_picker.server}
\title{Select cluster resolution}
\usage{
cluster_resolution_picker.ui(
  id,
  label = "Cluster resolutions",
  include_label_switch = FALSE
)

cluster_resolution_picker.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to select a cluster resolution to plot
}
\details{
Updates the Seurat object reactive values. Could use \code{SetIdent} instead?
}
\examples{

\dontrun{
ui <- fluidPage(cluster_resolution_picker.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(cluster_resolution_picker.server, id='page_name')
}}

}
