% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_environments.R
\name{collect_environments}
\alias{collect_environments}
\title{Collect environments for modules}
\usage{
collect_environments(module, id)
}
\arguments{
\item{module}{name of the module used to determine module-specific environment to load}
}
\value{
none
}
\description{
Creates environments for module-specific and seuratvis-wide variables
}
\details{
assigns a \code{seuratvis_env} and \code{module_env} environment in the calling function, including copying the \code{id} variable that is passed to \code{callModule()}
}
