% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\docType{data}
\name{colour_palette}
\alias{colour_palette}
\title{Initialise the colour palette in the namespace}
\format{An object of class \code{character} of length 40.}
\usage{
colour_palette
}
\description{
Initialise the colour palette in the namespace
}
\keyword{datasets}
