% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_output.dataset_info.r
\name{dataset_info_text_box.defaults}
\alias{dataset_info_text_box.defaults}
\title{Requires font-awesome version 5+ to be installed in shiny/www/shared/fontawesome}
\usage{
dataset_info_text_box.defaults(...)
}
\description{
Requires font-awesome version 5+ to be installed in shiny/www/shared/fontawesome
}
