% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.density_plot.R
\name{density_plot.ui}
\alias{density_plot.ui}
\alias{density_plot.server}
\title{Make a density plot of a feature}
\usage{
density_plot.ui(id, feature)

density_plot.server(input, output, session, seurat, cell_filtering, ...)
}
\arguments{
\item{id}{unique name of the element}

\item{feature}{name of the column in the Seurat object to plot}
}
\description{
Creates a density plot and thresholds
}
\examples{

\dontrun{
ui <- fluidPage(density_plot.ui(id='page_name', feature='column_name'))
server <- function(input, output, session) {
  callModule(density_plot.server, id='page_name-column_name')
}} 

}
