% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.elbow_plot.R
\name{elbow_plot.ui}
\alias{elbow_plot.ui}
\alias{elbow_plot.server}
\title{Make an elbow plot of a reduction}
\usage{
elbow_plot.ui(id)

elbow_plot.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a plot of points and thresholds
}
