% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.feature_ridges.r
\name{feature_ridge_by_idents.server}
\alias{feature_ridge_by_idents.server}
\title{Plot ridges for a feature across all idents}
\usage{
feature_ridge_by_idents.server(
  input,
  output,
  session,
  picked_feature,
  picked_clusters
)
}
\description{
Plot ridges for a feature across all idents
}
