% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.feature_ridges.r
\name{feature_ridges_by_ident.server}
\alias{feature_ridges_by_ident.server}
\title{Plot ridges for multiple features in an ident class}
\usage{
feature_ridges_by_ident.server(
  input,
  output,
  session,
  picked_feature,
  picked_clusters
)
}
\description{
Plot ridges for multiple features in an ident class
}
