% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.feature_values_per_cluster_plot.R
\name{feature_values_per_cluster_plot.ui}
\alias{feature_values_per_cluster_plot.ui}
\alias{feature_values_per_cluster_plot.server}
\title{Make a plot of a feature values in clusters}
\usage{
feature_values_per_cluster_plot.ui(id)

feature_values_per_cluster_plot.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a plot to show median and range of selected feature values in cluster idents
}
\examples{

\dontrun{
ui <- fluidPage(feature_values_per_cluster.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(feature_values_per_cluster.server, id='page_name')
}} 

}
