% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.features_per_cell_filter.R
\name{features_per_cell_filter.ui}
\alias{features_per_cell_filter.ui}
\alias{features_per_cell_filter.server}
\title{Specify unique features per cell limits}
\usage{
features_per_cell_filter.ui(
  id,
  label = "Features per cell",
  low = TRUE,
  high = TRUE
)

features_per_cell_filter.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}

\item{low, high}{include inputs for lower and upper thresholds, defaults to min/max of Seurat object}
}
\description{
Provides a way to define limits on number of features identified in a cell
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(features_per_cell_filter.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(features_per_cell_filter.server, id='page_name')
}}

}
