% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{filtering_parameters.reactions}
\alias{filtering_parameters.reactions}
\title{Reactive lists for dataset filetering}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
filtering_parameters.reactions
}
\description{
Reactive lists for dataset filetering
}
\keyword{datasets}
