% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_seurat_objects.R
\name{find_seurat_objects}
\alias{find_seurat_objects}
\title{Find Seurat objects}
\usage{
find_seurat_objects()
}
\value{
A \code{data.frame}
}
\description{
Look for Seurat objects in environments
}
\details{
The function uses \code{ls()} to get to first get a list of environments in the global environment. For each environment, \code{ls()} is again used to list the variables. If a variable is a \code{Seurat} object (its \code{class} is \code{Seurat}), it is retained in the output \code{data.frame}.
}
