% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gene_modules_plot.R
\name{gene_module_score_plot.ui}
\alias{gene_module_score_plot.ui}
\alias{gene_module_score_plot.server}
\title{Make a gene module score density plot}
\usage{
gene_module_score_plot.ui(id)

gene_module_score_plot.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a density plot of a gene module score
}
