% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.gene_name_and_description.R
\name{gene_name_and_description_text_box.ui}
\alias{gene_name_and_description_text_box.ui}
\alias{gene_name_and_description_text_box.server}
\title{Display gene information}
\usage{
gene_name_and_description_text_box.ui(id, width = 12)

gene_name_and_description_text_box.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{input, oputput, session}{used internally}
}
\description{
Displays gene of interest and biomaRt description
}
\examples{

\dontrun{
ui <- fluidPage(gene_name_and_description_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(gene_name_and_description_text_box.server(id='page_name'))}

}
