% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formatted_gene_description.R
\name{get_formatted_gene_description}
\alias{get_formatted_gene_description}
\title{Get description of a gene}
\usage{
get_formatted_gene_description(external_gene_name, mart)
}
\arguments{
\item{mart}{}
}
\value{
a character string of the \code{attribute} from the \code{mart} for the \code{external_gene_name}.
}
\description{
Use biomaRt to get formatted description of a gene from gene name
}
