% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jackstraw_plot.R
\name{jackstraw_plot.ui}
\alias{jackstraw_plot.ui}
\alias{jackstraw_plot.server}
\title{Make a JackStraw plot of a reduction}
\usage{
jackstraw_plot.ui(id)

jackstraw_plot.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a JackStraw plot
}
