% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jackstraw_pvalue_plot.R
\name{jackstraw_pvalue_plot.ui}
\alias{jackstraw_pvalue_plot.ui}
\alias{jackstraw_pvalue_plot.server}
\title{Make an JackStraw pvalue plot of a reduction}
\usage{
jackstraw_pvalue_plot.ui(id)

jackstraw_pvalue_plot.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a plot of points and thresholds
}
