% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knee_plot.R
\name{knee_plot.ui}
\alias{knee_plot.ui}
\alias{knee_plot.server}
\title{Make a knee plot of a feature}
\usage{
knee_plot.ui(id, feature)

knee_plot.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{feature}{name of the column in the Seurat object to plot}
}
\description{
Creates a plot of points and thresholds
}
\examples{

\dontrun{
ui <- fluidPage(knee_plot.ui(id='page_name', feature='column_name'))
server <- function(input, output, session) {
  callModule(knee_plot.server, id='page_name-column_name')
}} 

}
