% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.median_reads_per_cell.R
\name{number_of_reads_per_cell_text_box.ui}
\alias{number_of_reads_per_cell_text_box.ui}
\alias{number_of_reads_per_cell_text_box.server}
\title{Display median number of reads per cell}
\usage{
number_of_reads_per_cell_text_box.ui(
  id,
  width = 12,
  subtitle = "default",
  f = median,
  f_name = deparse(substitute(f))
)

number_of_reads_per_cell_text_box.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{subtitle}{type of subtitle to display; one of: \code{default}}

\item{f, f_name}{summary function taking one vector and returning a value and a name used to make the subtitle}

\item{input, oputput, session}{used internally}
}
\description{
Show the median number of reads in cells of the dataset
}
\examples{

\dontrun{
ui <- fluidPage(number_of_reads_per_cell_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(number_of_reads_per_cell_text_box.server, id='page_name')}

}
