% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.number_of_cells.R
\name{number_of_cells_text_box.ui}
\alias{number_of_cells_text_box.ui}
\alias{number_of_cells_text_box.server}
\title{Display number of cells in a dataset}
\usage{
number_of_cells_text_box.ui(id, width = 12)

number_of_cells_text_box.server(
  input,
  output,
  session,
  seurat,
  cell_filtering,
  ...
)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{input, output, session}{used internally}
}
\description{
Show the total number of cells
}
\examples{

\dontrun{
ui <- fluidPage(number_of_cells_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(number_of_cells_text_box.server, id='page_name')}

}
