% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.number_of_clusters.R
\name{number_of_clusters_text_box.ui}
\alias{number_of_clusters_text_box.ui}
\alias{number_of_clusters_text_box.server}
\title{Display number of clusters in a dataset}
\usage{
number_of_clusters_text_box.ui(id, width = 12, subtitle = "default")

number_of_clusters_text_box.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{subtitle}{type of subtitle to display; one of: \code{default}}

\item{input, oputput, session}{used internally}
}
\description{
Show the total number of clusters in a map
}
\examples{

\dontrun{
ui <- fluidPage(number_of_clusters_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(number_of_clusters_text_box.server, id='page_name')}

}
