% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.opacity_slider.R
\name{opacity_slider.ui}
\alias{opacity_slider.ui}
\alias{opacity_slider.server}
\title{Set an opacity}
\usage{
opacity_slider.ui(id, label = "Opacity")

opacity_slider.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to set an opacity reactive value
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(opacity_slider.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(opacity_slider.server, id='page_name')
}}

}
