% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.parse_ns_label.R
\name{parse_ns_label}
\alias{parse_ns_label}
\title{Parse NS label}
\usage{
parse_ns_label(x)
}
\description{
Remove the interim namespace label
}
\details{
Removes the intermediate namespace tag where the same module is use in multiple parts of the same tab. For example \code{tab-module-id} becomes \code{tab-id}.
}
