% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pca_heatmap.R
\name{pca_heatmap.ui}
\alias{pca_heatmap.ui}
\alias{pca_heatmap.server}
\title{Make a PCA heatmap of a reduction}
\usage{
pca_heatmap.ui(id)

pca_heatmap.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Creates a heatmap of features contributing to components
}
