% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.gene_name_and_description.R
\name{picked_feature_and_description_text_box.ui}
\alias{picked_feature_and_description_text_box.ui}
\alias{picked_feature_and_description_text_box.server}
\title{Display gene information}
\usage{
picked_feature_and_description_text_box.ui(id, width = 12)

picked_feature_and_description_text_box.server(
  input,
  output,
  session,
  seurat,
  ...
)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{input, output, session}{used internally}
}
\description{
Displays gene of interest and biomaRt description
}
\examples{

\dontrun{
ui <- fluidPage(picked_feature_and_description_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(picked_feature_and_description_text_box.server, id='page_name')}

}
