% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{plotting_options.rv}
\alias{plotting_options.rv}
\title{Plotting reactive values}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
plotting_options.rv
}
\description{
Plotting reactive values
}
\keyword{datasets}
