% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.point_size_slider.R
\name{point_size_slider.ui}
\alias{point_size_slider.ui}
\alias{point_size_slider.server}
\title{Set point sizes}
\usage{
point_size_slider.ui(id, label = "Point size")

point_size_slider.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to set a point size reactive value
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(point_size_slider.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(point_size_slider.server, id='page_name')
}}

}
