% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferred_choice.R
\name{preferred_choice}
\alias{preferred_choice}
\title{Select element of vecto by preference}
\usage{
preferred_choice(x, preferences, default = 1)
}
\arguments{
\item{x}{vector from which to choose}

\item{preferences}{ranked vector of terms to serach for in \code{x}}

\item{default}{if no \code{preferences} are found, the \code{default} element of \code{x} is returned}
}
\value{
an element of \code{x}
}
\description{
Returns the element of a vector using a ranked preference list
}
