% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.component_selector.R
\name{principle_component_picker.ui}
\alias{principle_component_picker.ui}
\alias{principle_component_picker.server}
\title{Select a principle component from a dataset PCA reduction}
\usage{
principle_component_picker.ui(
  id,
  label = "Principle component selection",
  include_picker = TRUE,
  include_slider = TRUE
)

principle_component_picker.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to select principle component(s) from a PCA reduction
}
\details{
Updates the Seurat object reactive values.
}
\examples{

}
