% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.project_name.R
\name{project_name_text_box.ui}
\alias{project_name_text_box.ui}
\alias{project_name_text_box.server}
\title{Display project name}
\usage{
project_name_text_box.ui(id, width = 12)

project_name_text_box.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{width}{integer value for width (rows sum to 12)}

\item{input, oputput, session}{used internally}
}
\description{
Create a formatted project name text box
}
\examples{

\dontrun{
ui <- fluidPage(project_name_text_box.ui(id='page_name'))
server <- function(input, output, session) callModule(project_name_text_box.server(id='page_name'))}

}
