% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_output.provenance_step_viewer.R
\name{provenance_step_viewer.ui}
\alias{provenance_step_viewer.ui}
\alias{provenance_step_viewer.server}
\title{Displya R script used for analysis}
\usage{
provenance_step_viewer.ui(id)

provenance_step_viewer.server(input, output, session, seurat, ...)
}
\arguments{
\item{id}{unique name of the ekement}
}
\description{
Provides a selector and formatted code display of scripts
}
\details{
Picks analysis step from the \code{misc@provenance} slot of the Seurat object.
}
