% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reduced_dimension_plot.R
\name{reduced_dimension_plot.ui}
\alias{reduced_dimension_plot.ui}
\title{Make a reduced dimension plot coloured by a feature}
\usage{
reduced_dimension_plot.ui(id, feature)
}
\arguments{
\item{id}{unique name of the element}

\item{feature}{either \code{selected_feature_resolution} or \code{picked_feature_values}}
}
\description{
Creates a scatterplot of a reduced dimensionality map
}
\examples{

\dontrun{
ui <- fluidPage(reduced_dimension_plot.ui(id='page_name', feature='feature'))
server <- function(input, output, session) {
  callModule(reduced_dimension_plot.server, id='page_name-feature')
}} 

}
