% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reduced_dimension_plot.R
\name{reduced_dimension_plot.server}
\alias{reduced_dimension_plot.server}
\title{Produce the ggplot object for a scatterplot}
\usage{
reduced_dimension_plot.server(input, output, session, seurat, ...)
}
\description{
Produce the ggplot object for a scatterplot
}
