% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{reference_metrics.rv}
\alias{reference_metrics.rv}
\title{Reactive list of summary statistics of loaded seurat object}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
reference_metrics.rv
}
\description{
Reactive list of summary statistics of loaded seurat object
}
\keyword{datasets}
