% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{selections.rv}
\alias{selections.rv}
\title{Reactive list of UI element selections}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
selections.rv
}
\arguments{
\item{`-picked_feature`}{name of picked feature from a tab}
}
\description{
Reactive list of UI element selections
}
\details{
Elements of the list can be accessed with \code{session$ns('feature')}.
}
\keyword{datasets}
