% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_output.show_filtering_parameters.R
\name{show_filtering_parameters.server}
\alias{show_filtering_parameters.server}
\title{React to filtering parameters}
\usage{
show_filtering_parameters.server(input, output, session)
}
\description{
React to filtering parameters
}
\details{
Updates the filtering parameters reactive values.
}
