% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_output.show_filtering_parameters.R
\name{show_filtering_parameters.ui}
\alias{show_filtering_parameters.ui}
\title{Print specified filtering parameters}
\usage{
show_filtering_parameters.ui(
  id,
  label = "Cell filtering parameters",
  include_copy_buttons = TRUE
)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}

\item{include_copy_buttons}{should buttons to copy formatted text be included?}
}
\description{
Displays the currently applied filters and can provide formatted copy buttons
}
\examples{

\dontrun{
ui <- fluidPage(show_filtering_parameters.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(show_filtering_parameters.server, id='page_name')
}}

#' @rdname total_umi_per_cell_filter

}
