% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.total_umi_per_cell_filter.R
\name{total_umi_per_cell_filter.ui}
\alias{total_umi_per_cell_filter.ui}
\alias{total_umi_per_cell_filter.server}
\title{Specify total UMIs per cell limits}
\usage{
total_umi_per_cell_filter.ui(
  id,
  label = "UMIs per cell",
  low = TRUE,
  high = TRUE
)

total_umi_per_cell_filter.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}

\item{low, high}{include inputs for lower and upper thresholds, defaults to min/max of Seurat object}
}
\description{
Provides a way to define limits on total UMI per cell
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(total_umi_per_cell_filter.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(total_umi_per_cell_filter.server, id='page_name')
}} 

}
