% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{ui_element_ids.env}
\alias{ui_element_ids.env}
\title{UI elements by module}
\format{An object of class \code{environment} of length 0.}
\usage{
ui_element_ids.env
}
\description{
UI elements by module
}
\keyword{datasets}
