% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.available_seurats.R
\name{available_seurats.ui}
\alias{available_seurats.ui}
\alias{available_seurats.server}
\alias{load_a_seurat.server}
\title{Display a table of available Seurat objects}
\usage{
available_seurats.ui(id)

available_seurats.server(input, output, session)

load_a_seurat.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}
}
\description{
Shows information on Seurats in accessible environments for selection
}
\details{
A \code{datatable} is rendered with additional Seurat metadata. The selected row is loaded into the viewer.
}
\examples{

\dontrun{
ui <- fluidPage(available_seurats.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(module=available_seurats.server, id='page_name')
  callModule(module=load_a_seurat.server, id='page_name')
}}

}
