% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_colour_palettes.R
\name{colour_palette.ui}
\alias{colour_palette.ui}
\alias{update_palette_type.server}
\alias{add_to_colour_palette.server}
\alias{colour_palette}
\title{Display a colour selection panel}
\usage{
colour_palette.ui(id, selectors = list(), include_full = FALSE)

update_palette_type.server(input, output, session)

add_to_colour_palette.server(input, output, session)

colour_palette(x)
}
\arguments{
\item{id}{unique name of the element}

\item{selectors}{list of lists with arguments to \code{colourInput}; requires \code{inputId}, \code{label} and best include \code{value} too}

\item{include_full}{boolean whether to create a square/limited palette switch}
}
\description{
Create UI elements for colours and a palette type switch

This is probably not a good place to keep this
}
\details{
For every selector in the list, a new element is created in the same order as provided. If a palette type switch is required, it is created after the selectors.
}
\examples{

\dontrun{
ui <- fluidPage(number_of_clusters_text_box.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(module=update_palette_type.server, id='page_name')
  callModule(module=add_to_colour_palette.server, id='page_name')
}}

}
