% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.feature_selector.R
\name{feature_picker.ui}
\alias{feature_picker.ui}
\alias{feature_picker.server}
\title{Select a feature from a dataset}
\usage{
feature_picker.ui(
  id,
  label = "Feature selection",
  include_metadata_switch = TRUE
)

feature_picker.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a method to select genes/features of a dataset
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(feature_picker.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(feature_picker.server, id='page_name')
}}

}
