% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.percent_mt_per_cell_filter.R
\name{percent_mt_per_cell_filter.ui}
\alias{percent_mt_per_cell_filter.ui}
\alias{percent_mt_per_cell_filter.server}
\title{Specify maximum mitochondrial proportion}
\usage{
percent_mt_per_cell_filter.ui(id, label = "Proportion Mt")

percent_mt_per_cell_filter.server(
  input,
  output,
  session,
  seurat,
  cell_filtering,
  ...
)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a way to define limit on proportion of mitochondrial UMI per cell
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(percent_mt_per_cell_filter.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(percent_mt_per_cell_filter.server, id='page_name')
}}

}
