% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.reduction_picker.R
\name{reduction_method_picker.ui}
\alias{reduction_method_picker.ui}
\alias{reduction_method_picker.server}
\title{Select reduction method}
\usage{
reduction_method_picker.ui(id, label = "Reduction method")

reduction_method_picker.server(input, output, session)
}
\arguments{
\item{id}{unique name of the element}

\item{label}{text label of the element}
}
\description{
Provides a way to select a reduction method to plot
}
\details{
Updates the Seurat object reactive values.
}
\examples{

\dontrun{
ui <- fluidPage(reduction_method_picker.ui(id='page_name'))
server <- function(input, output, session) {
  callModule(reduction_method_picker.server, id='page_name')
}}

}
